pyproject.toml
# The pyproject.toml file：https://python-poetry.org/docs/pyproject/

[project]
name = "example"
dynamic = ["version"]
description = ""
authors = [
  { name = "duanluan", email = "duanluan@example.com" }
]
license = "MIT"
readme = "README.md"

[tool.poetry]
version = "0.1.0"
packages = [
  { include = "common", from = "src" },
  { include = "app1", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.9,<=3.13"
pyyaml = "^6.0.3"
py7zr = "^1.0.0"
apscheduler = "^3.11.0"
pyinstaller = "^6.16.0"
pydantic="^2.11.9"
pydantic-settings="^2.11.0"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app1 = "app1.app1:main"
